+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/DeveloperOptionsFragment.kt
@@ -1,9 +1,10 @@
 package com.woocommerce.android.ui.prefs
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
+import androidx.fragment.app.viewModels
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.woocommerce.android.R
 import com.woocommerce.android.databinding.FragmentDeveloperOptionsBinding
 import com.woocommerce.android.ui.base.BaseFragment
@@ -11,23 +12,30 @@ import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class DeveloperOptionsFragment : BaseFragment(R.layout.fragment_developer_options) {
-    private var _binding: FragmentDeveloperOptionsBinding? = null
-    private val binding get() = _binding!!
-    // private val viewModel: DeveloperOptionsViewModel by viewModels()
+    val viewModel: DeveloperOptionsViewModel by viewModels()
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentDeveloperOptionsBinding.inflate(inflater, container, false)
-        return binding.root
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val binding = FragmentDeveloperOptionsBinding.bind(view)
+
+        initViews(binding)
+        observeViewState(binding)
+    }
+
+    private fun initViews(binding: FragmentDeveloperOptionsBinding) {
+        binding.developerOptionsRv.layoutManager = LinearLayoutManager(requireContext())
+        binding.developerOptionsRv.addItemDecoration(
+            DividerItemDecoration(requireContext(), DividerItemDecoration.VERTICAL)
+        )
+        binding.developerOptionsRv.adapter = DeveloperOptionsAdapter()
+    }
+
+    private fun observeViewState(binding: FragmentDeveloperOptionsBinding) {
+        viewModel.viewState.observe(viewLifecycleOwner) { state ->
+            (binding.developerOptionsRv.adapter as DeveloperOptionsAdapter).setItems(state.rows)
+        }
     }
 
     override fun getFragmentTitle() = resources.getString(R.string.dev_options)
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
 }