+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/sitepicker/SitePickerViewModel.kt
@@ -15,7 +15,7 @@ import com.woocommerce.android.analytics.ExperimentTracker
 import com.woocommerce.android.experiment.JetpackTimeoutExperiment
 import com.woocommerce.android.extensions.getSiteName
 import com.woocommerce.android.extensions.isSimpleWPComSite
-import com.woocommerce.android.support.HelpActivity
+import com.woocommerce.android.support.help.HelpActivity
 import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.ui.common.UserEligibilityFetcher
 import com.woocommerce.android.ui.login.AccountRepository
@@ -28,6 +28,7 @@ import com.woocommerce.android.ui.sitepicker.SitePickerViewModel.SitesListItem.N
 import com.woocommerce.android.ui.sitepicker.SitePickerViewModel.SitesListItem.WooSiteUiModel
 import com.woocommerce.android.util.FeatureFlag
 import com.woocommerce.android.util.WooLog
+import com.woocommerce.android.util.WooLog.T
 import com.woocommerce.android.viewmodel.LiveDataDelegate
 import com.woocommerce.android.viewmodel.MultiLiveEvent
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.Exit
@@ -38,8 +39,10 @@ import com.woocommerce.android.viewmodel.ResourceProvider
 import com.woocommerce.android.viewmodel.ScopedViewModel
 import com.woocommerce.android.viewmodel.navArgs
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import kotlinx.parcelize.Parcelize
 import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooErrorType
@@ -67,6 +70,9 @@ class SitePickerViewModel @Inject constructor(
     companion object {
         private const val WOOCOMMERCE_INSTALLATION_URL = "https://wordpress.com/plugins/woocommerce/"
         private const val WOOCOMMERCE_INSTALLATION_DONE_URL = "marketplace/thank-you/woocommerce"
+        private const val WOOCOMMERCE_STORE_CREATION_URL = "https://woocommerce.com/start"
+        private const val WOOCOMMERCE_STORE_CREATION_DONE_URL = "calypso/images/wpcom-ecommerce"
+        private const val WOOCOMMERCE_STORE_CREATION_RETRY_INTERVAL = 2000L
     }
 
     private val navArgs: SitePickerFragmentArgs by savedState.navArgs()
@@ -78,7 +84,7 @@ class SitePickerViewModel @Inject constructor(
     val sites: LiveData<List<SitesListItem>> = _sites
 
     private var loginSiteAddress: String?
-        get() = savedState.get("key") ?: appPrefsWrapper.getLoginSiteAddress()
+        get() = savedState["key"] ?: appPrefsWrapper.getLoginSiteAddress()
         set(value) = savedState.set("key", value)
 
     val shouldShowToolbar: Boolean
@@ -90,12 +96,32 @@ class SitePickerViewModel @Inject constructor(
             false -> loadStorePickerView()
         }
         updateSiteViewDetails()
+        loadAndDisplayUserInfo()
         loadAndDisplaySites()
+        if (appPrefsWrapper.getIsNewSignUp()) startStoreCreationWebFlow()
+    }
+
+    private fun loadAndDisplayUserInfo() = launch {
+        fun getUserInfo() = accountRepository.getUserAccount()?.let {
+            UserInfo(displayName = it.displayName, username = it.userName ?: "", userAvatarUrl = it.avatarUrl)
+        }
+
+        if (accountRepository.getUserAccount() != null) {
+            sitePickerViewState = sitePickerViewState.copy(userInfo = getUserInfo())
+        } else {
+            accountRepository.fetchUserAccount().fold(
+                onSuccess = {
+                    sitePickerViewState = sitePickerViewState.copy(userInfo = getUserInfo())
+                },
+                onFailure = {
+                    triggerEvent(ShowSnackbar(string.error_fetch_my_profile))
+                }
+            )
+        }
     }
 
     private fun updateSiteViewDetails() {
         sitePickerViewState = sitePickerViewState.copy(
-            userInfo = getUserInfo(),
             primaryBtnText = resourceProvider.getString(string.continue_button),
             secondaryBtnText = resourceProvider.getString(string.login_try_another_account),
             currentSitePickerState = SitePickerState.StoreListState
@@ -112,11 +138,12 @@ class SitePickerViewModel @Inject constructor(
         }
     }
 
-    private suspend fun fetchSitesFromApi(showSkeleton: Boolean) {
+    private suspend fun fetchSitesFromApi(showSkeleton: Boolean, delayTime: Long = 0) {
         sitePickerViewState = sitePickerViewState.copy(
             isSkeletonViewVisible = showSkeleton
         )
 
+        delay(delayTime)
         val startTime = System.currentTimeMillis()
         val result = repository.fetchWooCommerceSites()
         val duration = System.currentTimeMillis() - startTime
@@ -213,7 +240,9 @@ class SitePickerViewModel @Inject constructor(
         }
         sitePickerViewState = sitePickerViewState.copy(
             hasConnectedStores = filteredSites.isNotEmpty(),
-            isPrimaryBtnVisible = wooSites.isNotEmpty()
+            isPrimaryBtnVisible = wooSites.isNotEmpty(),
+            isNoStoresViewVisible = false,
+            currentSitePickerState = SitePickerState.StoreListState
         )
         loginSiteAddress?.let { processLoginSiteAddress(it) }
     }
@@ -337,10 +366,6 @@ class SitePickerViewModel @Inject constructor(
         )
     }
 
-    private fun getUserInfo() = accountRepository.getUserAccount()?.let {
-        UserInfo(displayName = it.displayName, username = it.userName ?: "", userAvatarUrl = it.avatarUrl)
-    }
-
     fun onSiteSelected(siteModel: SiteModel) {
         val updatedSites = _sites.value?.map {
             when (it) {
@@ -402,6 +427,17 @@ class SitePickerViewModel @Inject constructor(
     fun onEnterSiteAddressClick() {
         analyticsTrackerWrapper.track(AnalyticsEvent.SITE_PICKER_ENTER_SITE_ADDRESS_TAPPED)
         triggerEvent(SitePickerEvent.NavigateToSiteAddressEvent)
+    }
+
+    fun onCreateSiteButtonClick() {
+        analyticsTrackerWrapper.track(AnalyticsEvent.SITE_PICKER_CREATE_SITE_TAPPED)
+        triggerEvent(
+            NavigateToWPComWebView(
+                url = WOOCOMMERCE_STORE_CREATION_URL,
+                validationUrl = WOOCOMMERCE_STORE_CREATION_DONE_URL,
+                title = resourceProvider.getString(string.create_new_store)
+            )
+        )
     }
 
     fun onTryAnotherAccountButtonClick() {
@@ -528,6 +564,42 @@ class SitePickerViewModel @Inject constructor(
         }
     }
 
+    fun onSiteCreated(urls: List<String>) {
+        launch {
+            sitePickerViewState = sitePickerViewState.copy(
+                isSkeletonViewVisible = false, isProgressDiaLogVisible = true
+            )
+
+            var newSite: SiteModel? = null
+            while (newSite?.hasWooCommerce != true) {
+                newSite = urls.firstNotNullOfOrNull { url -> repository.getSiteBySiteUrl(url) }
+                if (newSite != null && newSite.hasWooCommerce) {
+                    onSiteSelected(newSite)
+                    onContinueButtonClick(true)
+                    WooLog.d(T.LOGIN, "Found new site: ${newSite.url}")
+                } else {
+                    WooLog.d(T.LOGIN, "New site not found, retrying...")
+                    val result = fetchSitesAfterCreation()
+                    if (result.isFailure) {
+                        triggerEvent(ShowSnackbar(string.site_picker_error))
+                        break
+                    } else {
+                        displaySites(result.getOrDefault(emptyList()))
+                    }
+                }
+            }
+        }
+    }
+
+    private suspend fun fetchSitesAfterCreation(): Result<List<SiteModel>> {
+        val result = withContext(Dispatchers.Default) {
+            delay(WOOCOMMERCE_STORE_CREATION_RETRY_INTERVAL)
+            repository.fetchWooCommerceSites()
+        }
+        return if (result.isError) Result.failure(Exception(result.error.message))
+        else Result.success(result.model ?: emptyList())
+    }
+
     fun onWooInstalled() {
         suspend fun fetchSite(site: SiteModel, retries: Int = 0): Result<SiteModel> {
             delay(retries * TimeUnit.SECONDS.toMillis(2))
@@ -596,6 +668,17 @@ class SitePickerViewModel @Inject constructor(
         fetchSitesFromApi(showSkeleton = true)
     }
 
+    private fun startStoreCreationWebFlow() {
+        appPrefsWrapper.markAsNewSignUp(false)
+        triggerEvent(
+            NavigateToWPComWebView(
+                url = WOOCOMMERCE_STORE_CREATION_URL,
+                validationUrl = WOOCOMMERCE_STORE_CREATION_DONE_URL,
+                title = resourceProvider.getString(string.create_new_store)
+            )
+        )
+    }
+
     private fun trackLoginEvent(
         currentFlow: UnifiedLoginTracker.Flow? = null,
         currentStep: UnifiedLoginTracker.Step? = null,
@@ -658,7 +741,12 @@ class SitePickerViewModel @Inject constructor(
         object NavigateToNewToWooEvent : SitePickerEvent()
         object NavigateToSiteAddressEvent : SitePickerEvent()
         data class NavigateToHelpFragmentEvent(val origin: HelpActivity.Origin) : SitePickerEvent()
-        data class NavigateToWPComWebView(val url: String, val validationUrl: String) : SitePickerEvent()
+        data class NavigateToWPComWebView(
+            val url: String,
+            val validationUrl: String,
+            val title: String? = null
+        ) : SitePickerEvent()
+
         data class NavigateToAccountMismatchScreen(
             val primaryButton: AccountMismatchPrimaryButton,
             val siteUrl: String,