+++ WooCommerce/src/main/kotlin/com/woocommerce/android/util/WooPermissionUtils.kt
@@ -9,7 +9,6 @@ import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.net.Uri
-import android.os.Build
 import android.provider.Settings
 import androidx.activity.result.ActivityResultLauncher
 import androidx.core.app.ActivityCompat
@@ -39,10 +38,10 @@ object WooPermissionUtils {
     }
 
     fun hasBluetoothScanPermission(context: Context) =
-        androidROrLower() || context.checkIfPermissionGiven(BLUETOOTH_SCAN)
+        SystemVersionUtils.isAtMostR() || context.checkIfPermissionGiven(BLUETOOTH_SCAN)
 
     fun hasBluetoothConnectPermission(context: Context) =
-        androidROrLower() || context.checkIfPermissionGiven(BLUETOOTH_CONNECT)
+        SystemVersionUtils.isAtMostR() || context.checkIfPermissionGiven(BLUETOOTH_CONNECT)
 
     fun requestFineLocationPermission(requestPermissionLauncher: ActivityResultLauncher<String>) {
         requestPermissionLauncher.launch(ACCESS_FINE_LOCATION)
@@ -60,6 +59,4 @@ object WooPermissionUtils {
 
     private fun Context.checkIfPermissionGiven(permission: String) =
         ContextCompat.checkSelfPermission(this, permission) == PackageManager.PERMISSION_GRANTED
-
-    private fun androidROrLower() = Build.VERSION.SDK_INT <= Build.VERSION_CODES.R
 }