+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/orders/OrderListViewModelTest.kt
@@ -10,6 +10,7 @@ import com.woocommerce.android.analytics.AnalyticsTrackerWrapper
 import com.woocommerce.android.extensions.NotificationReceivedEvent
 import com.woocommerce.android.extensions.takeIfNotEqualTo
 import com.woocommerce.android.model.RequestResult
+import com.woocommerce.android.model.UiString
 import com.woocommerce.android.push.NotificationChannelType
 import com.woocommerce.android.tools.NetworkStatus
 import com.woocommerce.android.tools.SelectedSite
@@ -28,6 +29,7 @@ import com.woocommerce.android.ui.orders.list.OrderListViewModel.OrderListEvent.
 import com.woocommerce.android.ui.payments.banner.BannerDisplayEligibilityChecker
 import com.woocommerce.android.ui.payments.banner.BannerState
 import com.woocommerce.android.util.LandscapeChecker
+import com.woocommerce.android.util.UtmProvider
 import com.woocommerce.android.util.getOrAwaitValue
 import com.woocommerce.android.util.observeForTesting
 import com.woocommerce.android.viewmodel.BaseUnitTest
@@ -94,6 +96,7 @@ class OrderListViewModelTest : BaseUnitTest() {
     private val bannerDisplayEligibilityChecker: BannerDisplayEligibilityChecker = mock()
     private val analyticsTrackerWrapper: AnalyticsTrackerWrapper = mock()
     private val landscapeChecker: LandscapeChecker = mock()
+    private val orderListUtmProvider: UtmProvider = mock()
 
     @Before
     fun setup() = testBlocking {
@@ -131,6 +134,7 @@ class OrderListViewModelTest : BaseUnitTest() {
             orderListTransactionLauncher = mock(),
             analyticsTrackerWrapper = analyticsTrackerWrapper,
             landscapeChecker = landscapeChecker,
+            orderListUtmProvider = orderListUtmProvider,
         )
     }
 
@@ -460,6 +464,9 @@ class OrderListViewModelTest : BaseUnitTest() {
             ).thenReturn(
                 "${AppUrls.WOOCOMMERCE_PURCHASE_CARD_READER_IN_COUNTRY}US"
             )
+            whenever(orderListUtmProvider.getUrlWithUtmParams(any())).thenReturn(
+                "${AppUrls.WOOCOMMERCE_PURCHASE_CARD_READER_IN_COUNTRY}US"
+            )
             whenever(
                 bannerDisplayEligibilityChecker.canShowCardReaderUpsellBanner(anyLong())
             ).thenReturn(true)
@@ -731,7 +738,13 @@ class OrderListViewModelTest : BaseUnitTest() {
             // WHEN
             val title = viewModel.bannerState.value?.title
 
-            assertThat(title).isEqualTo(R.string.card_reader_upsell_card_reader_banner_title)
+            assertThat(title).isEqualTo(
+                UiString.UiStringRes(
+                    stringRes = R.string.card_reader_upsell_card_reader_banner_title,
+                    params = emptyList(),
+                    containsHtml = false
+                )
+            )
         }
     }
 
@@ -747,7 +760,13 @@ class OrderListViewModelTest : BaseUnitTest() {
             // WHEN
             val description = viewModel.bannerState.value?.description
 
-            assertThat(description).isEqualTo(R.string.card_reader_upsell_card_reader_banner_description)
+            assertThat(description).isEqualTo(
+                UiString.UiStringRes(
+                    stringRes = R.string.card_reader_upsell_card_reader_banner_description,
+                    params = emptyList(),
+                    containsHtml = false
+                )
+            )
         }
     }
 
@@ -763,7 +782,13 @@ class OrderListViewModelTest : BaseUnitTest() {
             // WHEN
             val primaryActionLabel = viewModel.bannerState.value?.primaryActionLabel
 
-            assertThat(primaryActionLabel).isEqualTo(R.string.card_reader_upsell_card_reader_banner_cta)
+            assertThat(primaryActionLabel).isEqualTo(
+                UiString.UiStringRes(
+                    stringRes = R.string.card_reader_upsell_card_reader_banner_cta,
+                    params = emptyList(),
+                    containsHtml = false
+                )
+            )
         }
     }
 
@@ -779,7 +804,13 @@ class OrderListViewModelTest : BaseUnitTest() {
             // WHEN
             val chipLabel = viewModel.bannerState.value?.chipLabel
 
-            assertThat(chipLabel).isEqualTo(R.string.card_reader_upsell_card_reader_banner_new)
+            assertThat(chipLabel).isEqualTo(
+                UiString.UiStringRes(
+                    stringRes = R.string.card_reader_upsell_card_reader_banner_new,
+                    params = emptyList(),
+                    containsHtml = false
+                )
+            )
         }
     }