+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 <!--
 Contains editorialized release notes. Raw release notes should go into `RELEASE-NOTES.txt`.
 -->
+## 11.0
+
+Good news! We added a new analytics section to the app! Simply click on the See More button under the store stats to check more detailed information on Revenue and Orders. This isn’t the last of it – more updates to come in the next few weeks!
+
+## 10.9
+
+This release includes a lot more stability fixes to help you log in much easier. You can now connect your WordPress.com account with Jetpack and access your store that much faster. Keep that feedback coming!
+
+## 10.8
+
+This release includes a handful of minor improvements and bug fixes that our team found. We are working hard on a few exciting features as well – and you’ll see that later this year. Thanks for using WooCommerce!
 
 ## 10.7